add not member status
