+++ main.py
@@ -2,12 +2,12 @@
 # -*- coding: utf-8 -*-
 
 from flask import Flask
-from flask import request
 from flask import json
+from flask import request
 
-import service as s
 import config
-import multithread_utils
+import service as s
+from utils import multithread_utils
 
 token = config.token
 confirmation_token = config.confirmation_token
@@ -57,7 +57,6 @@ def main(argv):
     port = int(argv[1])
     updater = multithread_utils.Thread_allow_updater()
     updater.start()
-    s.send_message_admins_after_restart()
     app.run(host='0.0.0.0', port=port, debug=False)
 
 if __name__ == '__main__':
